refactor: main 클래스로 이동
