+++ src/main/java/study/RacingInfoProvider.java
@@ -1,0 +1,9 @@
+package study;
+
+import java.util.Set;
+
+interface RacingInfoProvider {
+    Set<Car> getCars();
+
+    int countSteps();
+}
