+++ src/main/java/study/ResultView.java
