+++ src/test/java/study/CarRacingTest.java
@@ -92,38 +92,7 @@ public class CarRacingTest {
         this.carRacing = new CarRacing(new StaticInfoProvider(laps, cars), resultView);
     }
 
-    private static class CarRacing {
-
-        private final Set<Car> cars;
-        private final int steps;
-        private final ResultView resultView;
-
-        public CarRacing(RacingInfoProvider racingInfoProvider, ResultView resultView) {
-            cars = racingInfoProvider.getCars();
-            steps = racingInfoProvider.countSteps();
-            this.resultView = resultView;
-        }
-
-        public void start() {
-            if (cars.isEmpty() || steps == 0)
-                throw new IllegalStateException("레이싱 정보가 존재하지 않습니다.");
-
-            for (int i = 0; i < steps; i++) {
-                move();
-            }
-        }
-
-        private void move() {
-            Set<LapResult> result = new HashSet<>();
-            for (Car car : cars) {
-                car.move();
-                result.add(new LapResult(car.getId(), car.isMoved()));
-            }
-            resultView.add(result);
-        }
-    }
-
-    private static class StaticInfoProvider implements RacingInfoProvider {
+    static class StaticInfoProvider implements RacingInfoProvider {
         private final int steps;
         private final Set<Car> cars;
 
@@ -141,12 +110,6 @@ public class CarRacingTest {
         public int countSteps() {
             return steps;
         }
-    }
-
-    private interface RacingInfoProvider {
-        Set<Car> getCars();
-
-        int countSteps();
     }
 
     private static class TestingCar implements Car {
