+++ src/test/java/study/StaticInfoProvider.java
@@ -1,0 +1,7 @@
+package study;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+
